% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eststore.R
\name{eststore}
\alias{eststore}
\title{JJF eststore function}
\usage{
eststore(est, estlist_arg = estlist)
}
\arguments{
\item{est}{the regression estimate to be stored}

\item{estlist_arg}{the list of stored regressions}
}
\value{
side effect style update the estlist_arg with the est
}
\description{
Store regression output in a list list statas eststore
}
\examples{
mtcars \%>\%
  lm(data = ., mpg ~ disp) \%>\%
  eststore(car_regs)

mtcars \%>\%
  lm(data = ., mpg ~ hp) \%>\%
  eststore(car_regs)

mtcars \%>\%
  lm(data = ., mpg ~ disp + hp) \%>\%
  eststore(car_regs)

car_regs[[1]] \%>\% summary()
car_regs[[2]] \%>\% summary()
car_regs[[3]] \%>\% summary()

# stargazer::stargazer(car_regs, type = "text")

}
\keyword{eststore}
