% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest_estout.R
\name{fixest_estout}
\alias{fixest_estout}
\title{Fixest to Tex Fragment Helper Function}
\usage{
fixest_estout(est_tex, file)
}
\arguments{
\item{est_tex}{tex output from etable with regressions to be texed out}

\item{file}{path to location to save tex output}
}
\value{
tex output
}
\description{
Helper function that will output a tex fragment
}
