% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest_nclusters.R
\name{fixest_nclusters}
\alias{fixest_nclusters}
\title{Number of Clusters Helper Function for Fixest}
\usage{
fixest_nclusters(est)
}
\arguments{
\item{est}{output from fixest::feols or another regression function}
}
\value{
row of number of clusters
}
\description{
Helper function that will output the number of clusters
}
