% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest_ymean.R
\name{fixest_ymean}
\alias{fixest_ymean}
\title{Ymeans Helper Function for Fixest}
\usage{
fixest_ymean(est)
}
\arguments{
\item{est}{output from fixest::feols or another regression function}
}
\value{
row of ymeans
}
\description{
Helper function that will output the ymean
}
