% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.historical.lookups.R
\docType{data}
\name{state.confederacy}
\alias{state.confederacy}
\title{State Historical Look Ups}
\format{An object of class \code{"vector"}}
\source{
Wikipedia, mostly
}
\usage{
data(state.confederacy)
}
\description{
Several simple vectors to use like state.abb or state.name
from the base datasets
Includes some for the Civil War, 13 colonies, union admission, etc
}
\examples{
x <- c("New York", "Virginia", "West Virginia", "Colorado")
state.confederacy[match(x, state.name)]
state.union[match(x, state.name)]
state.civilwar[match(x, state.name)]
state.admission[match(x, state.name)]
state.13.colony[match(x, state.name)]
state.slavery.1860[match(x, state.name)]
}
\keyword{datasets}
