% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.fips.R
\docType{data}
\name{state.fips}
\alias{state.fips}
\title{State FIPS look up}
\format{
An object of class \code{"vector"}
}
\source{
\href{https://gist.github.com/soodoku/f9e18efe98f7d74931d8b4a79a49e6f5}{Sooduko}
}
\usage{
data(state.fips)
}
\description{
Simple vector to use like state.abb or state.name
from the base datasets
This has fips codes for states
}
\examples{
data(state.fips)
x <- c("New York", "Virginia")
state.fips[match(x, state.name)]
}
\keyword{datasets}
