% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.historical.lookups.R
\docType{data}
\name{state.historical}
\alias{state.historical}
\alias{state.confederacy}
\alias{state.union}
\alias{state.civilwar}
\alias{state.admission}
\alias{state.13.colony}
\alias{state.slavery.1860}
\title{State Historical Look Ups}
\format{
An object of class \code{"vector"}

An object of class \code{logical} of length 50.

An object of class \code{logical} of length 50.

An object of class \code{character} of length 50.

An object of class \code{character} of length 50.

An object of class \code{logical} of length 50.

An object of class \code{logical} of length 50.
}
\source{
Wikipedia, mostly
}
\usage{
data(state.confederacy)

data(state.union)

data(state.civilwar)

data(state.admission)

data(state.13.colony)

data(state.slavery.1860)

state.confederacy

state.union

state.civilwar

state.admission

state.13.colony

state.slavery.1860
}
\description{
Several simple vectors to use like state.abb or state.name
from the base datasets
Includes some for the Civil War, 13 colonies, union admission, etc
}
\examples{
x <- c("New York", "Virginia", "West Virginia", "Colorado")
state.confederacy[match(x, state.name)]
# state in joined the Confederacy?
state.union[match(x, state.name)]
# state remained loyal to Union?
state.civilwar[match(x, state.name)]
# state's side in the Civil War (Union, Confederacy, or Territory)
state.admission[match(x, state.name)]
# year of state's admission to the US
state.13.colony[match(x, state.name)]
# an original 13 colony state?
state.slavery.1860[match(x, state.name)]
# did the state permit slavery in 1860?

}
\keyword{datasets}
