% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.icpsr.R
\docType{data}
\name{state.icpsr}
\alias{state.icpsr}
\title{State ICPSR look up}
\format{
An object of class \code{"vector"}
}
\source{
\href{https://gist.github.com/soodoku/4c6bbf529fab6ec0bc3576e829af7c5d}{Sooduko}
}
\usage{
data(state.icpsr)
}
\description{
Simple vector to use like state.abb or state.name
from the base datasets
This has icpsr codes
}
\examples{
data(state.icpsr)
x <- c("New York", "Virginia")
state.icpsr[match(x, state.name)]
}
\keyword{datasets}
