% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.ipums.bpl.R
\docType{data}
\name{state.ipums.bpl}
\alias{state.ipums.bpl}
\title{State IPUMS BPL look up}
\format{An object of class \code{"vector"}}
\source{
\href{https://usa.ipums.org/usa-action/variables/BPL#codes_section}{IPUMS BPL Codes}
}
\usage{
data(state.ipums.bpl)
}
\description{
Simple vector to use like state.abb or state.name
from the base datasets
This has bpl codes from ipums
}
\examples{
data(state.ipums.bpl)
x <- c("New York", "Virginia")
state.ipums.bpl[match(x, state.name)]
}
\keyword{datasets}
