% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_to_cong.R
\name{year_to_cong}
\alias{year_to_cong}
\title{JJF Convert between Congress Number and Election Year}
\usage{
year_to_cong(year)
}
\arguments{
\item{year}{the election year (1788 to 2018, roughly)}
}
\description{
Some data is by year and some with congress number
}
\examples{

# the 100th congress was elected in 1986

year_to_cong(1986)

}
