% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cong_to_year.R
\name{cong_to_year}
\alias{cong_to_year}
\title{Convert between Congress Number and Election Year}
\usage{
cong_to_year(cong)
}
\arguments{
\item{cong}{the congress number (1 to... 115 as of 2018, but any integer will work)}
}
\description{
Some data is by year and some with congress number.
    Inevitably, you need it the other way...
}
\examples{

# the 100th congress was elected in 1986

cong_to_year(100)

}
