% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipums_codes.R
\name{ipums_codes}
\alias{ipums_codes}
\title{JJF scrape ipums codes}
\usage{
ipums_codes(variable, detailed = FALSE)
}
\arguments{
\item{variable}{the name of the variable from ipums we want to scrape}

\item{detailed}{get the general or detailed codes default FALSE for just general}
}
\description{
Pull the IPUMS codes to variable label lookup from the IPUMS website
}
\examples{
\dontrun{

# get bpl codes

bpl_codes_1 <- ipums_codes("bpl", detailed = TRUE)

bpl_codes_2 <- ipums_codes("bpl", detailed = FALSE)

}
}
