% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jjf.R
\name{theme_jjf}
\alias{theme_jjf}
\title{JJF Theme Function}
\usage{
theme_jjf(font = "Roboto Condensed")
}
\arguments{
\item{font}{the font to use in the graph with Roboto Condensed as the default}
}
\description{
Add my custom ggplot2 theme
}
\examples{
ggplot(cars, aes(x = dist, y = speed)) +
  geom_point() +
  theme_jjf()

# same as above, roboto Condensed is the default
ggplot(cars, aes(x = dist, y = speed)) +
  geom_point() +
  theme_jjf(font = "Roboto Condensed")

ggplot(cars, aes(x = dist, y = speed)) +
  geom_point() +
  theme_jjf(font = "CMU Sans Serif")

}
\keyword{theme}
